backend/api/apps/organizations/serializers.py
from rest_framework import serializers
from .models import (
    Organization,
    Position,
)


class OrganizationSerializer(serializers.ModelSerializer):
    class Meta:
        model = Organization
        fields = "__all__"


class PositionSerializer(serializers.ModelSerializer):
    class Meta:
        model = Position
        fields = "__all__"
