backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.1",
    "django-cors-headers>=4.7.0",
    "django-environ>=0.12.0",
    "djangorestframework>=3.16.0",
    "djangorestframework-simplejwt>=5.5.0",
    "psycopg2-binary>=2.9.10",
    "pytils>=0.4.3",
]

[dependency-groups]
dev = [
    "django-debug-toolbar>=5.2.0",
    "pytest>=8.4.0",
    "pytest-django>=4.11.1",
]
